Move create intersection events use case from plugin to application layer
