+++ OTAnalytics/application/analysis/intersect.py
@@ -9,6 +9,10 @@ from OTAnalytics.domain.track import TrackId
 from OTAnalytics.domain.types import EventType
 
 
+class IntersectionError(Exception):
+    pass
+
+
 class RunIntersect(ABC):
     """
     Interface defining the use case to intersect the given tracks with the given
